+++ include/BDSOutputROOTEventHeader.hh
@@ -23,6 +23,7 @@ along with BDSIM.  If not, see <http://www.gnu.org/licenses/>.
 #include "TObject.h"
 
 #include <string>
+#include <vector>
 
 /** 
  * @brief Information about the software and the file.
@@ -45,6 +46,8 @@ public:
   std::string fileType;
   int         dataVersion;           ///< Our data format version.
   bool        doublePrecisionOutput; ///< Whether using double precision output - assumed float if not
+  std::vector<std::string> analysedFiles; ///< List of which files were analysed in case of a rebdsim output file.
+  std::vector<std::string> combinedFiles; ///< List of which files were combined in case of a rebdsimCombine output file.
 
   /// Update the file type.
   void SetFileType(std::string fileTypeIn) {fileType = fileTypeIn;}
@@ -53,9 +56,10 @@ public:
   /// Nominally, we don't expose the fill methods to the analysis root dictionaries
   /// but as this doesn't use geant4 and is required when creating analysis output
   /// file, we break that convention.
-  void Fill();
+  void Fill(const std::vector<std::string>& analysedFilesIn = std::vector<std::string>(),
+	    const std::vector<std::string>& combinedFilesIn = std::vector<std::string>());
 
-  ClassDef(BDSOutputROOTEventHeader,2);
+  ClassDef(BDSOutputROOTEventHeader,3);
 };
 
 #endif