+++ src/BDSOutputROOTEventInfo.cc
@@ -25,8 +25,8 @@ ClassImp(BDSOutputROOTEventInfo)
 BDSOutputROOTEventInfo::BDSOutputROOTEventInfo():
   startTime(time_t()),
   stopTime(time_t()),
-  duration(0),
-  cpuTime(std::clock_t()),
+  durationWall(0),
+  durationCPU(0),
   seedStateAtStart(""),
   index(-1),
   aborted(false),
@@ -52,8 +52,8 @@ void BDSOutputROOTEventInfo::Flush()
 {
   startTime         = time_t();
   stopTime          = time_t();
-  duration          = 0;
-  cpuTime           = std::clock_t();
+  durationWall      = 0;
+  durationCPU       = 0;
   seedStateAtStart  = "";
   index             = -1;
   aborted           = false;
@@ -76,23 +76,23 @@ void BDSOutputROOTEventInfo::Fill(const BDSOutputROOTEventInfo* other)
   if (!other)
     {return;}
 
-  startTime              = other->startTime;
-  stopTime               = other->stopTime;
-  duration               = other->duration;
-  cpuTime                = other->cpuTime;
-  seedStateAtStart       = other->seedStateAtStart;
-  index                  = other->index;
-  aborted                = other->aborted;
-  primaryHitMachine      = other->primaryHitMachine;
+  startTime               = other->startTime;
+  stopTime                = other->stopTime;
+  durationWall            = other->durationWall;
+  durationCPU             = other->durationCPU;
+  seedStateAtStart        = other->seedStateAtStart;
+  index                   = other->index;
+  aborted                 = other->aborted;
+  primaryHitMachine       = other->primaryHitMachine;
   primaryAbsorbedInCollimator = other->primaryAbsorbedInCollimator;
-  memoryUsageMb          = other->memoryUsageMb;
-  energyDeposited        = other->energyDeposited;
-  energyDepositedVacuum  = other->energyDepositedVacuum;
-  energyDepositedWorld   = other->energyDepositedWorld;
+  memoryUsageMb           = other->memoryUsageMb;
+  energyDeposited         = other->energyDeposited;
+  energyDepositedVacuum   = other->energyDepositedVacuum;
+  energyDepositedWorld    = other->energyDepositedWorld;
   energyDepositedWorldContents = other->energyDepositedWorldContents;
-  energyDepositedTunnel  = other->energyDepositedTunnel;
+  energyDepositedTunnel   = other->energyDepositedTunnel;
   energyImpactingAperture = other->energyImpactingAperture;
-  energyKilled           = other->energyKilled;
-  energyTotal            = other->energyTotal;
-  nCollimatorsInteracted = other->nCollimatorsInteracted;
+  energyKilled            = other->energyKilled;
+  energyTotal             = other->energyTotal;
+  nCollimatorsInteracted  = other->nCollimatorsInteracted;
 }